Add S3 Bucket policy
Policy allows List and GetObject on the bucket only for
IAM Role "Udagram-EC2-S3-ReadOnly". Web servers should
be assigned this role at creation to allow getting the
application from the bucket.

This role is also assigned to the JumpBox to allow it to
retrieve the security key to be used with the servers
located on the private subnet.