Configure JumpBox launch config
Change AMI to Amazon Linux. As it is only used to
SSH into other servers, the overhead of Ubuntu server
is not required.

The UserData script will get the security key for the
private subnet servers, put it in the ec2-user home dir
and set the ec2-user as the owner. It will also set the
permissions to the requirements by AWS. This allows an
admin to simply SSH into the JumpBox and directly SSH
from there into the private servers without requiring
any configuration.